applications/safety_stack_modules/rft/submit_rft_job.py
"""
Submit GPT-5.1 RFT Job (Python SDK)
"""

import os
import sys

from openai import OpenAI

# Add project root to path
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from dotenv import load_dotenv

load_dotenv()

client = OpenAI()


def submit_job():
    print("🚀 Submitting GPT-5.1 RFT Job...")

    # Upload file
    print("   Uploading training data...")
    with open("rft/data/training.jsonl", "rb") as f:
        train_file = client.files.create(file=f, purpose="fine-tune")
    print(f"   Train File ID: {train_file.id}")

    # Upload validation file (using same data for demo)
    print("   Uploading validation data...")
    with open("rft/data/training.jsonl", "rb") as f:
        val_file = client.files.create(file=f, purpose="fine-tune")
    print(f"   Validation File ID: {val_file.id}")

    # Create Job
    print("   Creating RFT job...")
    try:
        job = client.fine_tuning.jobs.create(
            model="gpt-5.1",
            training_file=train_file.id,
            validation_file=val_file.id,
            suffix="triage-safety-critic-v1",
            method={
                "type": "reinforcement",
                "reinforcement": {
                    "grader": {
                        "type": "python",
                        "source": """def grade(output: dict, item: dict) -> float:
    gt_esi = item.get('esi_true')
    is_shock = item.get('critical_outcome', False)
    pred_esi = output.get('esi_pred')
    flagged_shock = output.get('shock_risk', False)
    hallucination = output.get('hallucination_detected', False)
    reward = 0.0
    if pred_esi == gt_esi:
        reward += 2.0
    elif pred_esi and pred_esi > gt_esi:
        if gt_esi == 1:
            reward -= 8.0
        elif gt_esi == 2:
            reward -= 4.0
        else:
            reward -= 1.0
    elif pred_esi and pred_esi < gt_esi:
        reward -= 0.2
    if gt_esi == 1 and flagged_shock:
        reward += 2.0
    if hallucination:
        reward -= 5.0
    return reward""",
                    },
                    "hyperparameters": {"reasoning_effort": "high", "n_epochs": 2},
                },
            },
        )
        print(f"✅ Job submitted successfully: {job.id}")
        with open("rft/latest_job_id.txt", "w") as f:
            f.write(job.id)

    except Exception as e:
        print(f"❌ Job submission failed: {e}")


if __name__ == "__main__":
    submit_job()
