scripts/tests_standalone/test_hallucination_safety.py
#!/usr/bin/env python3
"""
Comprehensive test for hallucination detection with timeout handling
"""

import asyncio
import os
import sys

sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from eval.hallucinations import detect_hallucinations


async def test_timeout_handling():
    """Test that hallucination detection handles timeouts gracefully."""
    print("=" * 60)
    print("TEST 1: Normal hallucination detection")
    print("=" * 60)

    case = {
        "age": 45,
        "sex": "M",
        "vital_signs": {"hr": 110, "sbp": 95},
        "chief_complaint": "Chest pain",
    }

    council_output = "Patient presents with chest pain. HR is 110.  BP is 95 systolic."

    result = await detect_hallucinations(case, council_output, timeout_seconds=10)

    assert "total_claims" in result, "Missing total_claims"
    assert "hallucination_rate" in result, "Missing hallucination_rate"
    assert "timed_out" in result, "Missing timed_out flag"

    print(f"✅ Total claims: {result['total_claims']}")
    print(f"✅ Hallucination rate: {result['hallucination_rate']:.1%}")
    print(f"✅ Timed out: {result['timed_out']}")
    print(f"✅ Error: {result.get('error', 'None')}")

    print("\n" + "=" * 60)
    print("TEST 2: Timeout handling (short timeout)")
    print("=" * 60)

    # Test with very short timeout to force timeout
    result2 = await detect_hallucinations(case, council_output, timeout_seconds=0.001)

    assert result2["timed_out"] == True, "Should have timed out"
    assert result2["error"] == "timeout", "Should have timeout error"
    assert result2["hallucination_rate"] == 0.0, "Should default to 0% on timeout"

    print("✅ Correctly handled timeout")
    print("✅ Defaulted to safe state (0% hallucinations)")

    print("\n" + "=" * 60)
    print("ALL TESTS PASSED ✅")
    print("=" * 60)


if __name__ == "__main__":
    asyncio.run(test_timeout_handling())
