src/tic/contract.py
"""
Monotonic Safety Contract (MSC) Loading and Validation
======================================================

Provides contract loading from YAML with JSON Schema validation.
"""

import json
from dataclasses import dataclass, field
from pathlib import Path
from typing import Any, Dict, List, Optional, Set

import yaml

# Path to the MSC JSON Schema
SCHEMA_PATH = Path(__file__).parent.parent.parent / "configs" / "schemas" / "msc.schema.json"


@dataclass
class EventDefinition:
    """Definition of a canonical event."""

    event_id: str
    description: str
    category: str = "ambiguous"
    clinical_significance: str = ""
    extraction_hints: List[str] = field(default_factory=list)


@dataclass
class AppliesTo:
    """Specification of which scenarios a contract applies to."""

    scenario_classes: List[str]
    description: str = ""
    exclusions: List[str] = field(default_factory=list)
    esi_levels: List[int] = field(default_factory=list)

    def matches(self, scenario_class: str, esi_level: Optional[int] = None) -> bool:
        """Check if this contract applies to a given scenario."""
        # Check exclusions first
        if scenario_class in self.exclusions:
            return False

        # Check scenario class
        if scenario_class not in self.scenario_classes:
            return False

        # Check ESI level if specified
        if esi_level is not None and self.esi_levels:
            if esi_level not in self.esi_levels:
                return False

        return True


@dataclass
class EntryRule:
    """Rule for entering a state."""

    when_any: List[str] = field(default_factory=list)
    when_all: List[str] = field(default_factory=list)
    from_states: List[str] = field(default_factory=list)

    def triggers(self, events: Set[str], current_state: str) -> bool:
        """Check if this rule triggers given events and current state."""
        # Check from_states constraint
        if self.from_states and current_state not in self.from_states:
            return False

        # Check when_any (OR condition)
        if self.when_any:
            if any(e in events for e in self.when_any):
                return True

        # Check when_all (AND condition)
        if self.when_all:
            if all(e in events for e in self.when_all):
                return True

        return False


@dataclass
class StateInvariant:
    """Invariant constraints for a state."""

    forbid_any: List[str] = field(default_factory=list)
    forbid_all: List[str] = field(default_factory=list)
    require_any: List[str] = field(default_factory=list)
    require_all: List[str] = field(default_factory=list)
    allow_any: List[str] = field(default_factory=list)
    description: str = ""

    def check_violation(self, events: Set[str]) -> Optional[Dict[str, Any]]:
        """
        Check if events violate this invariant.

        Returns None if no violation, or a dict with violation details.
        """
        # Check forbid_any (violation if ANY forbidden event occurs)
        for event in self.forbid_any:
            if event in events:
                return {
                    "type": "forbid_any",
                    "violated_constraint": "forbid_any",
                    "triggering_event": event,
                    "events_present": list(events),
                }

        # Check forbid_all (violation if ALL forbidden events occur)
        if self.forbid_all:
            if all(e in events for e in self.forbid_all):
                return {
                    "type": "forbid_all",
                    "violated_constraint": "forbid_all",
                    "triggering_events": self.forbid_all,
                    "events_present": list(events),
                }

        # Check require_any (violation if NONE of required events occur)
        if self.require_any:
            if not any(e in events for e in self.require_any):
                return {
                    "type": "require_any",
                    "violated_constraint": "require_any",
                    "missing_events": self.require_any,
                    "events_present": list(events),
                }

        # Check require_all (violation if ANY required event is missing)
        if self.require_all:
            missing = [e for e in self.require_all if e not in events]
            if missing:
                return {
                    "type": "require_all",
                    "violated_constraint": "require_all",
                    "missing_events": missing,
                    "events_present": list(events),
                }

        return None


@dataclass
class MonotonicitySpec:
    """Monotonicity constraints."""

    irreversible_states: List[str]
    description: str = ""


@dataclass
class ContractMetadata:
    """Contract metadata."""

    author: str = ""
    created: str = ""
    last_modified: str = ""
    clinical_basis: str = ""
    reviewed_by: List[str] = field(default_factory=list)


@dataclass
class MonotonicSafetyContract:
    """
    Monotonic Safety Contract (MSC).

    Defines trajectory-level safety invariants for a class of scenarios.
    """

    contract_id: str
    version: str
    description: str
    applies_to: AppliesTo
    states: List[str]
    events: List[EventDefinition]
    entry_rules: Dict[str, EntryRule]
    monotonicity: MonotonicitySpec
    invariants: Dict[str, StateInvariant]
    metadata: ContractMetadata = field(default_factory=ContractMetadata)

    @property
    def initial_state(self) -> str:
        """Get the initial state (first in list)."""
        return self.states[0]

    def get_event_definition(self, event_id: str) -> Optional[EventDefinition]:
        """Get event definition by ID."""
        for event in self.events:
            if event.event_id == event_id:
                return event
        return None

    def is_irreversible(self, state: str) -> bool:
        """Check if a state is irreversible (one-way latch)."""
        return state in self.monotonicity.irreversible_states

    def get_next_state(self, current_state: str, events: Set[str]) -> str:
        """
        Determine next state given current state and events.

        Respects monotonicity constraints (irreversible states).
        """
        # If current state is irreversible, stay in it
        if self.is_irreversible(current_state):
            return current_state

        # Check entry rules for each state
        for state, rule in self.entry_rules.items():
            if rule.triggers(events, current_state):
                return state

        # No transition, stay in current state
        return current_state

    def check_invariant(self, state: str, events: Set[str]) -> Optional[Dict[str, Any]]:
        """
        Check if events violate the invariant for a state.

        Returns None if no violation, or violation details.
        """
        if state not in self.invariants:
            return None

        return self.invariants[state].check_violation(events)

    def applies_to_scenario(self, scenario_class: str, esi_level: Optional[int] = None) -> bool:
        """Check if this contract applies to a scenario."""
        return self.applies_to.matches(scenario_class, esi_level)


def load_contract(path: Path) -> MonotonicSafetyContract:
    """
    Load a contract from a YAML file.

    Args:
        path: Path to the YAML contract file

    Returns:
        Loaded MonotonicSafetyContract

    Raises:
        FileNotFoundError: If file doesn't exist
        yaml.YAMLError: If YAML is invalid
        ValueError: If contract structure is invalid
    """
    with open(path, "r") as f:
        data = yaml.safe_load(f)

    return _parse_contract(data)


def _parse_contract(data: Dict[str, Any]) -> MonotonicSafetyContract:
    """Parse contract data into a MonotonicSafetyContract object."""

    # Parse applies_to
    applies_to_data = data.get("applies_to", {})
    applies_to = AppliesTo(
        scenario_classes=applies_to_data.get("scenario_classes", []),
        description=applies_to_data.get("description", ""),
        exclusions=applies_to_data.get("exclusions", []),
        esi_levels=applies_to_data.get("esi_levels", []),
    )

    # Parse events
    events = []
    for event_data in data.get("events", []):
        events.append(
            EventDefinition(
                event_id=event_data["event_id"],
                description=event_data.get("description", ""),
                category=event_data.get("category", "ambiguous"),
                clinical_significance=event_data.get("clinical_significance", ""),
                extraction_hints=event_data.get("extraction_hints", []),
            )
        )

    # Parse entry rules
    entry_rules = {}
    for state, rule_data in data.get("entry_rules", {}).items():
        entry_rules[state] = EntryRule(
            when_any=rule_data.get("when_any", []),
            when_all=rule_data.get("when_all", []),
            from_states=rule_data.get("from_states", []),
        )

    # Parse monotonicity
    mono_data = data.get("monotonicity", {})
    monotonicity = MonotonicitySpec(
        irreversible_states=mono_data.get("irreversible_states", []),
        description=mono_data.get("description", ""),
    )

    # Parse invariants
    invariants = {}
    for state, inv_data in data.get("invariants", {}).items():
        invariants[state] = StateInvariant(
            forbid_any=inv_data.get("forbid_any", []),
            forbid_all=inv_data.get("forbid_all", []),
            require_any=inv_data.get("require_any", []),
            require_all=inv_data.get("require_all", []),
            allow_any=inv_data.get("allow_any", []),
            description=inv_data.get("description", ""),
        )

    # Parse metadata
    meta_data = data.get("metadata", {})
    metadata = ContractMetadata(
        author=meta_data.get("author", ""),
        created=meta_data.get("created", ""),
        last_modified=meta_data.get("last_modified", ""),
        clinical_basis=meta_data.get("clinical_basis", ""),
        reviewed_by=meta_data.get("reviewed_by", []),
    )

    return MonotonicSafetyContract(
        contract_id=data["contract_id"],
        version=data["version"],
        description=data.get("description", ""),
        applies_to=applies_to,
        states=data["states"],
        events=events,
        entry_rules=entry_rules,
        monotonicity=monotonicity,
        invariants=invariants,
        metadata=metadata,
    )


def validate_contract(path: Path) -> Dict[str, Any]:
    """
    Validate a contract against the JSON Schema.

    Args:
        path: Path to the YAML contract file

    Returns:
        Dict with 'valid' bool and 'errors' list
    """
    try:
        import jsonschema
    except ImportError:
        return {
            "valid": False,
            "errors": ["jsonschema package not installed. Run: pip install jsonschema"],
        }

    # Load schema
    if not SCHEMA_PATH.exists():
        return {
            "valid": False,
            "errors": [f"Schema not found at {SCHEMA_PATH}"],
        }

    with open(SCHEMA_PATH, "r") as f:
        schema = json.load(f)

    # Load contract
    try:
        with open(path, "r") as f:
            contract_data = yaml.safe_load(f)
    except yaml.YAMLError as e:
        return {
            "valid": False,
            "errors": [f"Invalid YAML: {e}"],
        }

    # Validate
    validator = jsonschema.Draft7Validator(schema)
    errors = list(validator.iter_errors(contract_data))

    if errors:
        return {
            "valid": False,
            "errors": [f"{e.json_path}: {e.message}" for e in errors],
        }

    return {"valid": True, "errors": []}
