pyproject.toml
[project]
name = "scribegoat2"
version = "0.6.0"
description = "Evaluation framework for multi-turn safety persistence in healthcare AI"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.51.0",
    "anthropic>=0.34.0",
    "autogen-agentchat>=0.7.5",
    "autogen-ext[openai]>=0.7.5",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "jinja2>=3.1.6",  # Security: CVE-2025-27516, CVE-2024-56326
    "pyyaml>=6.0.1",
    "scikit-learn>=1.4.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "cryptography>=46.0.0",  # Security: CVE-2024-12797, CVE-2024-6119
]

[project.optional-dependencies]
# RAG dependencies (heavy, not required for core evaluation)
rag = [
    "sentence-transformers>=2.2.0",  # Dense embeddings
    "openem>=0.2.0",                 # OpenEM corpus retrieval
]
# Production RAG backends (even heavier)
rag-prod = [
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",              # FAISS vector search
]
# Analysis dependencies (pressure stratification regression etc.)
analysis = [
    "statsmodels>=0.14.0",
]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.5.0",
    "ruff>=0.1.9",
    "pre-commit>=3.6.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "detect-secrets>=1.4.0",
]

[project.scripts]
scribegoat-fhir = "src.fhir.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "evaluation"]

[tool.pytest.ini_options]
# Must match CI_PYTHONPATH in .github/workflows/tests.yml.
# Project root (.) first so root-level packages take precedence over archived duplicates.
pythonpath = [
    ".",
    "applications/council_architecture",
    "applications/safety_stack_modules",
    "research/archive/support_modules",
    "research/archive/old_evaluation_dirs",
    "research/archive/old_analysis",
    "research/archive",
    "scripts/runners",
    "scripts/analysis",
    "benchmarks",
]
testpaths = ["tests", "evaluation/evals"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: tests that require external API keys (OPENAI_API_KEY, ANTHROPIC_API_KEY)",
]
addopts = "-v --tb=short"
timeout = 300
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [
    "results",
    "spc_results",
    "output",
    "reports",
    "audit_logs",
    "tmp",
    "vendor",
    "data",
    "research",
    "bloom_medical_eval",
    "bloom_eval_v2",
    "evaluation126",
    "lir",
    "evaluation/bloom_eval_v2/results",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E402", "E501", "W291", "W293"]

[tool.ruff.lint.per-file-ignores]
# Re-exports in __init__.py files
"**/__init__.py" = ["F401"]
# Legacy modules — broad suppression for pre-existing patterns
"applications/**/*.py" = ["F401", "F841", "E722", "F821", "F402"]
"tools/**/*.py" = ["F841", "E722"]
"skills/**/*.py" = ["F841", "E712", "E722"]
"benchmarks/**/*.py" = ["F841", "E722", "F402"]
# Scripts — import guards and legacy patterns
"scripts/smoke_test_*.py" = ["F401", "F841"]
"scripts/analysis/*.py" = ["F401", "F841", "E722"]
"scripts/validate_contracts.py" = ["F401", "F841"]
"scripts/test_mcp_integration.py" = ["F841", "E712"]
"scripts/generate_observability_artifacts.py" = ["F841"]
"scripts/generate_integrity_report.py" = ["F841", "E722"]
"scripts/governance/*.py" = ["F841", "F402"]
"scripts/runners/*.py" = ["F841", "E722"]
"scripts/verify_claim_evidence.py" = ["F841"]
"scripts/demo_workflow.py" = ["F841"]
"scripts/replay_for_full_fidelity.py" = ["F841"]
"scripts/run_full_evaluation.py" = ["E722", "F841"]
"scripts/tests_standalone/*.py" = ["F841", "E722", "E712"]
# Tests — import guards and unused vars in test fixtures
"tests/test_constitutional_ai.py" = ["F401", "F841"]
"tests/test_confidence_intervals.py" = ["F401", "F841"]
"tests/test_model_comparator.py" = ["F841", "E712"]
"tests/test_ensemble_validator.py" = ["F841"]
"tests/test_persona_constraints.py" = ["F841", "E712", "E722"]
"tests/test_ci_validation.py" = ["F841"]
"tests/test_draft_contract_guard.py" = ["F841"]
"tests/test_dual_engine.py" = ["F841"]
"tests/test_fhir_generator.py" = ["F841"]
"tests/test_safety_corrections.py" = ["F841"]
"tests/test_abstention_gradients.py" = ["F841"]
"tests/skill_tests/frontier_lab_validation.py" = ["F841"]
# Evaluation harness
"evaluation/ensemble_evaluator.py" = ["F841"]
# Core source — targeted suppression
"src/mcp_server/server.py" = ["F821"]
"src/observability/exporter.py" = ["F841", "E741"]
"src/observability/instrumentation.py" = ["F841"]
"src/metrics/clinical_risk_profile.py" = ["F401"]
"src/metrics/multi_contract_profile.py" = ["F841"]
"src/fhir/generator.py" = ["F841"]
"src/tic/events.py" = ["F841"]
"src/tsr/verification/agents/adversarial.py" = ["E741"]
"src/tsr/verification/agents/clinical.py" = ["E741"]

[tool.coverage.run]
source = ["src", "evaluation"]
omit = ["**/test_*.py", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
